fix pg type conversions

+++ migrations/000001_initial-schema.up.sql
@@ -25,7 +25,7 @@ CREATE TABLE dish (
     id uuid DEFAULT uuid_generate_v4() PRIMARY KEY,
     created_at timestamp without time zone NOT NULL DEFAULT now(),
     name character varying(50) NOT NULL,
-    price bigint NOT NULL,
+    price int NOT NULL,
     description text NOT NULL,
     restaurant_id uuid REFERENCES restaurant(id)
 );
